growth-agent/requirements.txt
crewai>=0.80.0
crewai-tools>=0.14.0
anthropic>=0.40.0
httpx>=0.27.0
python-dotenv>=1.0.0
praw>=7.8.0
google-api-python-client>=2.150.0
truststore>=0.9.1
